feat(eventbridge)!: switch `rules` variable to type `map` (#32)
As part of the migration to `for_each`, we need to provide users with a means
of providing identifiers that are known at apply time. Allow users to provide
stable identifiers through the keys in a map.

BREAKING CHANGE: `rules` must be converted from a list to a map. Keys in this map are
only used for addressing purposes within terraform.